% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outfit.R
\docType{package}
\name{outfit}
\alias{outfit}
\title{outfit: Outlier Detection in Multivariate Functional Data}
\description{
Outlier Detection in Multivariate Functional Data
}
\note{
The name \code{outfit} comes from \code{OUTlier detection in multivariate FunctIonal DaTa}
}
\references{
Sodja, C. (2020). Detecting Anomalous Time Series by GAMLSS-Akaike-Weights-Scoring. arXiv preprint arXiv:2002.00499.

Talagala, P. D., Hyndman, R. J., Smith-Miles, K. (2020). Anomaly Detection in High Dimensional Data.
Journal of Computational and Graphical Statistics, (To appear).
}
\seealso{
The core functions in this package: \code{\link{get_features}}
}
